Implement clear_vibrancy via removing the blurred view from the window; implement a NSVisualEffectView subview with tag for matching the blurred view in the window; add IntelliJ IDEA and macOS-specific files to gitignore
