+++ .gitignore
@@ -4,3 +4,6 @@
 
 /target
 Cargo.lock
+.idea/
+*.iml
+.DS_Store
